Use GNU gettext for translations
Translations are now managed in gettext format (pot, po-files). During
building the RPM, they are compiled into mo-files, which are used by
Python and qm-files, which are used by QML. In QML we need to now use
qsTranslate instead of qsTr so that we can specify the context.

This means that we can avoid using the primitive and buggy Qt linguist
tools (apart from converting po to qm) and we can, in the future, if
needed, do things like string formatting in translatable strings in
Python and reading translatable strings from e.g. JSON or XML files in
Python.

As a downside, used converters probably can't handle plural forms or
contexts, so we'll refrain from using those. The context name in the
resulting qm-file is blank, so in QML we'll use qsTranslate("", "Foo");